pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "file-auto-pipeline"
version = "0.1.0"
description = "Manufacturing Docs Pipeline - 제조 검사 데이터를 보고서로 자동 변환"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Team", email = "team@example.com"}
]

dependencies = [
    # Core
    "pyyaml>=6.0",
    # Render
    "docxtpl>=0.16",
    "openpyxl>=3.1",
    # App
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8",
    "jinja2>=3.1",
    "httpx>=0.26",
    # AI Providers
    "anthropic>=0.18",
    "google-generativeai>=0.3",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "pytest-playwright>=0.4",
    "playwright>=1.40",
    "ruff>=0.2",
    "mypy>=1.8",
    "bandit>=1.7",
    "types-PyYAML>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in default argument (FastAPI File/Form pattern)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # assert 사용 허용
